requirements.txt
openai
rembg
Pillow
requests
tqdm
